requirements.txt
flask==2.3.3
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==20.6
plotly==5.17.0
dash==2.14.1
dash-bootstrap-components==1.5.0
sqlite3
python-dotenv==1.0.0
PyJWT==2.8.0
stripe==7.5.0
paypalrestsdk==1.13.3
bcrypt==4.0.1
cryptography==41.0.4
